Improved console output. * Moved console code and types to services. * Added new `logWithLevel` function to colour code output * Refined certain errors * Updated employer_sites to have better names
