+++ server/scripts/services/index.ts
@@ -1,1 +1,3 @@
 export { changeCohort } from './change-cohort';
+export * from './console';
+export * from './shared-types';
